+++ packages/bridge-controller/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Export `AmountsAndAsset` type
 - Add utils to support fee validation when a quote's fees are denominated in multiple assets
   - `hasSufficientGasForQuote` returns true if the wallet's balances are greater than or equal to the quote's network fees
+  - `hasNetworkFee` returns true if any of the network fee object's amounts are defined and non-zero
 
 ### Changed